Add Mog, Gogo, and Umaro to the world of ruin randomization
Changed randomization so Mog can be at Fanatics Tower or Relm can be at Umaro's cave, but not both at the same time, and Mog, Gogo, and Umaro can't be the character who comes free on the Falcon (because they need to have an event for naming).
Mog handles checking for whether you named him in the WoB at all locations.
The locations are included in the character section of the log.