Adjust guaranteed miab probability
For bsaib, re-add the crappy event enemies, because removing them threw the chance of early vanilla miabs way out of whack.
For default settings, make it less likely to pick one of the guaranteed miabs, since there are many fewer now, and adjust probability to keep the number of miabs about the same.